faster bw recovery
Now bwUL/bwDL will recover to maxUL/maxDL after at least 500 milliseconds.

Theoretically, this will help prevent CAKE from staying in 1 Mbit/s when the latency varies a lot.

This might need further testing for various network conditions to see if this commit solves the issue (i.e. if there is a lot of users using the DNS and causing the DNS latency to vary a lot).